refactor(perl-corpus): extract metadata and loading modules (#0000)
