+++ crates/perl-corpus/src/loading/mod.rs
@@ -1,0 +1,5 @@
+pub mod dir;
+pub mod file;
+
+pub use dir::parse_dir;
+pub use file::parse_file;
