+++ crates/perl-corpus/src/metadata/mod.rs
@@ -1,0 +1,10 @@
+mod flags;
+mod ids;
+pub mod parser;
+pub mod query;
+pub mod section;
+mod syntax;
+mod tags;
+
+pub use query::{find_by_flag, find_by_tag};
+pub use section::Section;
