+++ crates/perl-corpus/src/metadata/query.rs
@@ -1,0 +1,11 @@
+use crate::metadata::section::Section;
+
+/// Find sections by tag
+pub fn find_by_tag<'a>(sections: &'a [Section], tag: &str) -> Vec<&'a Section> {
+    sections.iter().filter(|s| s.has_tag(tag)).collect()
+}
+
+/// Find sections by flag
+pub fn find_by_flag<'a>(sections: &'a [Section], flag: &str) -> Vec<&'a Section> {
+    sections.iter().filter(|s| s.has_flag(flag)).collect()
+}
